Refactor redux structure
Separating actions from their respective reducer only introduced confusion when creating new ones
