+++ app/assets/scripts/redux/index.js
